% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_osha.R
\name{clean_osha_addresses}
\alias{clean_osha_addresses}
\title{Provide a Cleaned Copy of the OSHA dataset}
\usage{
clean_osha_addresses(osha)
}
\description{
In particular, this cleans up the address data by
splitting it out into columns for each major component.
}
