% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_states.R
\name{infer_states}
\alias{infer_states}
\title{Infer States for Open Cases}
\usage{
infer_states(osha)
}
\description{
This function matches the open complaints to the state in which the
complaint occured.
}
\details{
The method used is to take the closed complaints (which have address data)
and the OSHA office IDs and to create a lookup table from OSHA office ID =>
state.

Then using said lookup table, code the open complaints.
}
