% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_naics.R
\name{load_detailed_naics}
\alias{load_detailed_naics}
\title{Load Detailed NAICS}
\usage{
load_detailed_naics(export = T)
}
\description{
This file contains detailed descriptions for every 6 digit NAICS code
https://www.census.gov/eos/www/naics/downloadables/downloadables.html
2017 NAICS Index File \link{XLSX, 498KB}
}
\details{
This file has descriptions for 2- or 3-digit NAICS codes
2017 NAICS Descriptions \link{XLSX, 264KB}
}
