% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_naics.R
\name{load_naics}
\alias{load_naics}
\title{Load NAICS classification data}
\usage{
load_naics(export = T)
}
\description{
The NAICS codes were retrieved from
https://www.census.gov/eos/www/naics/
2017 NAICS Manual
}
