% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_osha.R
\name{make_osha_counts_w_county_data}
\alias{make_osha_counts_w_county_data}
\title{Provide OSHA Counts by County Map with Data Included}
\usage{
make_osha_counts_w_county_data(
  export = T,
  print_frc_matched = F,
  print_unmatched = F
)
}
\arguments{
\item{print_frc_matched}{A logical argument to indicate whether or not
to print the number of counties which were successfully matched of those
that weren't originally NA.}
}
\description{
Provide OSHA Counts by County Map with Data Included
}
